+++ examples/hackernews-clone/src/main.rs
@@ -7,7 +7,7 @@ use futures_signals::signal::{Mutable, SignalExt};
 use reqwasm::http::Request;
 use serde::{de::DeserializeOwned, Deserialize};
 use silkenweb::{
-    clone,
+    clone, css_classes,
     elements::html::{self, a, div, h1, h2, header, li, nav, p, span, ul, Div, Li, A},
     mount,
     node::element::{Element, ElementBuilder, ParentBuilder},
@@ -16,6 +16,8 @@ use silkenweb::{
     task::spawn_local,
 };
 use timeago::Formatter;
+
+css_classes!("hackernews.scss");
 
 type Result<T> = result::Result<T, reqwasm::Error>;
 
@@ -37,9 +39,9 @@ impl App {
 
     fn render(&self) -> Div {
         div()
-            .class(["page"])
+            .class([PAGE])
             .child(
-                header().child(nav().child(h1().class(["page-banner"]).children([
+                header().child(nav().child(h1().class([PAGE_BANNER]).children([
                     router::anchor("topstories").text("Top"),
                     router::anchor("newstories").text("New"),
                     router::anchor("askstories").text("Ask"),
@@ -48,7 +50,7 @@ impl App {
             )
             .child(
                 html::main()
-                    .class(["page-content"])
+                    .class([PAGE_CONTENT])
                     .child_signal(self.0.signal_ref(|content| {
                         if let Some(content) = content {
                             content.render()
@@ -148,12 +150,12 @@ impl Story {
 
         div()
             .child(
-                h2().class(["story-title"])
+                h2().class([STORY_TITLE])
                     .child(a().href(self.url.as_ref()).text(&self.title)),
             )
             .child(
                 span()
-                    .class(["de-emphasize"])
+                    .class([DE_EMPHASIZE])
                     .child(span().text(&format!("{score} point{} by ", plural(score))))
                     .child(user_link(&self.by))
                     .child(span().text(&format!(" {time_ago} | ")))
@@ -201,7 +203,7 @@ impl UserDetails {
             .child(text_as_html(&self.user.about))
             .child(span().text(&format!("{} karma", self.user.karma)))
             .child(
-                ul().class(["user-submitted-stories"])
+                ul().class([USER_SUBMITTED_STORIES])
                     .children(self.submitted.iter().map(Story::render)),
             )
             .build()
@@ -248,10 +250,10 @@ impl CommentTree {
 
     fn render(&self, depth: usize) -> Li {
         let time_ago = time_ago(self.comment.time);
-        li().class(["comment"])
+        li().class([COMMENT])
             .child(
                 span()
-                    .class(["de-emphasize"])
+                    .class([DE_EMPHASIZE])
                     .child(user_link(&self.comment.by))
                     .text(&format!(" {time_ago}")),
             )
@@ -270,7 +272,7 @@ impl CommentTree {
 fn text_as_html(text: &str) -> Div {
     let text = text.to_owned();
     div()
-        .class(["user-content"])
+        .class([USER_CONTENT])
         .effect(move |elem| elem.set_inner_html(&text))
         .build()
 }