+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Server side routing
 - Add `router::anchor` and `router::link_clicked` convenience methods
+- Compile time constants for CSS classes can be generated with `css_classes`
 
 ## 0.2.0 - 2021-02-09